+++ barcode/lib/src/upca.dart
@@ -82,6 +82,7 @@ class BarcodeUpcA extends BarcodeEan {
     double width,
     double height,
     double fontHeight,
+    double textPadding,
   ) {
     if (!drawText) {
       return 0;
@@ -96,6 +97,7 @@ class BarcodeUpcA extends BarcodeEan {
     double width,
     double height,
     double fontHeight,
+    double textPadding,
   ) {
     if (!drawText) {
       return 0;
@@ -111,16 +113,18 @@ class BarcodeUpcA extends BarcodeEan {
     double width,
     double height,
     double fontHeight,
+    double textPadding,
     bool drawText,
   ) {
     if (!drawText) {
-      return super.getHeight(index, count, width, height, fontHeight, drawText);
+      return super.getHeight(
+          index, count, width, height, fontHeight, textPadding, drawText);
     }
 
-    final h = height - fontHeight;
+    final h = height - fontHeight - textPadding;
 
     if (index + count < 11 || (index > 45 && index < 49) || index > 82) {
-      return h + fontHeight / 2;
+      return h + fontHeight / 2 + textPadding;
     }
 
     return h;
@@ -132,12 +136,13 @@ class BarcodeUpcA extends BarcodeEan {
     double width,
     double height,
     double fontHeight,
+    double textPadding,
     double lineWidth,
   ) sync* {
     final text = checkLength(data, maxLength);
     final w = lineWidth * 7;
-    final left = marginLeft(true, width, height, fontHeight);
-    final right = marginRight(true, width, height, fontHeight);
+    final left = marginLeft(true, width, height, fontHeight, textPadding);
+    final right = marginRight(true, width, height, fontHeight, textPadding);
 
     yield BarcodeText(
       left: 0,